refactor: add LocationsPopup, remove LocationPicker
