+++ src/components/crm/MedicalInfo/index.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import {
   Controller,
   useFieldArray,
